# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SwiftLingo - Telegram Mini App translation marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateMe
      tags: [users]
      summary: Update current user settings (e.g. preferred language)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: Not found
  /translator-profiles/me:
    get:
      operationId: getMyTranslatorProfile
      tags: [translatorProfiles]
      summary: Get current user's translator profile
      responses:
        "200":
          description: Translator profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranslatorProfile"
        "404":
          description: No translator profile
    put:
      operationId: upsertMyTranslatorProfile
      tags: [translatorProfiles]
      summary: Create or update current user's translator profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TranslatorProfileInput"
      responses:
        "200":
          description: Translator profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranslatorProfile"
  /translator-profiles/{userId}:
    get:
      operationId: getTranslatorProfile
      tags: [translatorProfiles]
      summary: Get a translator profile by user ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Translator profile with stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranslatorProfileWithStats"
        "404":
          description: Not found
  /translator-applications:
    post:
      operationId: submitTranslatorApplication
      tags: [translatorApplications]
      summary: Submit a translator application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TranslatorApplicationInput"
      responses:
        "201":
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranslatorApplication"
  /translator-applications/me:
    get:
      operationId: getMyTranslatorApplication
      tags: [translatorApplications]
      summary: Get current user's translator application
      responses:
        "200":
          description: Translator application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranslatorApplication"
        "404":
          description: No application found
...
  /reviews:
    post:
      operationId: submitReview
      tags: [reviews]
      summary: Client submits a review for completed contract
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Review submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /reviews/translator/{userId}:
    get:
      operationId: listTranslatorReviews
      tags: [reviews]
      summary: List reviews for a translator
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
...
components:
  schemas:
...
    User:
      type: object
      required: [id, telegramId, role, createdAt]
      properties:
        id:
          type: integer
        telegramId:
          type: string
        username:
          type: ["string", "null"]
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        profilePhotoUrl:
          type: ["string", "null"]
        role:
          type: string
          enum: [client, translator, admin]
        preferredLanguage:
          type: string
          enum: [en, uz, ru]
        isTranslatorApproved:
          type: boolean
        createdAt:
          type: string
    UserUpdate:
      type: object
      properties:
        preferredLanguage:
          type: string
          enum: [en, uz, ru]
        role:
          type: string
          enum: [client, translator]
    TranslatorProfile:
      type: object
      required: [id, userId, pricePerWord, sourceLanguages, targetLanguages, specializations, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        bio:
          type: ["string", "null"]
        certifications:
          type: ["string", "null"]
        yearsOfExperience:
          type: ["integer", "null"]
        pricePerWord:
          type: number
        sourceLanguages:
          type: array
          items:
            type: string
        targetLanguages:
          type: array
          items:
            type: string
        specializations:
          type: array
          items:
            type: string
        rating:
          type: ["number", "null"]
        completedJobsCount:
          type: integer
        createdAt:
          type: string
    TranslatorProfileWithStats:
      allOf:
        - $ref: "#/components/schemas/TranslatorProfile"
        - type: object
          properties:
            user:
              $ref: "#/components/schemas/User"
            recentReviews:
              type: array
              items:
                $ref: "#/components/schemas/Review"
    TranslatorProfileInput:
      type: object
      required: [pricePerWord, sourceLanguages, targetLanguages, specializations]
      properties:
        bio:
          type: string
        certifications:
          type: string
        yearsOfExperience:
          type: integer
        pricePerWord:
          type: number
        sourceLanguages:
          type: array
          items:
            type: string
        targetLanguages:
          type: array
          items:
            type: string
        specializations:
          type: array
          items:
            type: string
    TranslatorApplication:
      type: object
      required: [id, userId, status, termsAcceptedAt, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        fullName:
          type: string
        bio:
          type: ["string", "null"]
        certifications:
          type: ["string", "null"]
        yearsOfExperience:
          type: ["integer", "null"]
        pricePerWord:
          type: number
        sourceLanguages:
          type: array
          items:
            type: string
        targetLanguages:
          type: array
          items:
            type: string
        specializations:
          type: array
          items:
            type: string
        status:
          type: string
          enum: [draft, pending, approved, rejected]
        termsAcceptedAt:
          type: string
        adminNote:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    TranslatorApplicationInput:
      type: object
      required: [fullName, pricePerWord, sourceLanguages, targetLanguages, specializations, acceptTerms]
      properties:
        fullName:
          type: string
        bio:
          type: string
        certifications:
          type: string
        yearsOfExperience:
          type: integer
        pricePerWord:
          type: number
        sourceLanguages:
          type: array
          items:
            type: string
        targetLanguages:
          type: array
          items:
            type: string
        specializations:
          type: array
          items:
            type: string
        acceptTerms:
          type: boolean
...
    Review:
      type: object
      required: [id, contractId, clientId, translatorId, rating, createdAt]
      properties:
        id:
          type: integer
        contractId:
          type: integer
        clientId:
          type: integer
        translatorId:
          type: integer
        rating:
          type: integer
          minimum: 1
          maximum: 5
        feedback:
          type: ["string", "null"]
        client:
          $ref: "#/components/schemas/User"
        createdAt:
          type: string
    ReviewInput:
      type: object
      required: [contractId, rating]
      properties:
        contractId:
          type: integer
        rating:
          type: integer
          minimum: 1
          maximum: 5
        feedback:
          type: string
...